backend/requirements.txt
asgiref==3.4.1
bidict==0.21.3
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==4.3.1
h11==0.12.0
imageio==2.9.0
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
networkx==2.6.3
opencv-python==4.5.3.56
opencv-python-headless
openslide-python==1.1.2
pandas==1.3.3
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2021.1
PyWavelets==1.1.1
Shapely==1.7.1
simple-websocket==0.4.0
six==1.16.0
tifffile==2021.8.30
typing-extensions==3.10.0.2
waitress==2.0.0
Werkzeug==2.0.1
wsproto==1.0.0
zipp==3.5.0
